Move subviews into new file
